Make process_pipe_input() handle json log
Changes process_pipe_input() to consider T/F to be either csv or json log
rather than just csv. Also change the actual write function in write_syslogger_file()
to handle the destination parameter being either value set via a |.

NOTE: Review whether this is a good idea as may make more sense to have the pipe
input itself be more intelligent and specify the log target rather than blindly
sending output to "one or the other" and having the csvlogFile or jsonlogFile
being defined dictate which actually gets written.

Perhaps have something like "c|C" for true/false for CSV, "j|J" for JSON, and
"t|f" (as before) for the neither case.

Otherwise there's a race condition if a message gets sent for one logger (ex: csv)
but the logging target is changed (ex: to json) in between the message being piped
and the log collector processing it.