Add PGHOSTCMD to libpq for ProxyCommand-style commands as socket I/O
Adds support for a PGHOSTCMD environment variable that, if specified, will be executed
with $SHELL as a child process to connect to the database rather than creating a TCP
or Unix socket. The standard input and output of the child will be treated as the output
and input of the socket to the database.

Currently only for Linux. Not tested on Windows or with multiple hosts.

Some examples:

Executing necat to open a socket:

    PGHOSTCMD='nc -q db.example.com:12345' psql

Executing ssh to tunnel through an SSH server:

    PGHOSTCMD='ssh bastion.example.com -- nc -q0 db.example.com 5432' psql

Executing ssh to tunnel to a remote server and connect to a local unix socket:

    PGHOSTCMD='ssh db.example.com -- nc -q0 -U /localhost:5432' psql